internal/database/code_monitor_emails.go
pbckbge dbtbbbse

import (
	"context"
	"dbtbbbse/sql"
	"time"

	"github.com/keegbncsmith/sqlf"

	"github.com/sourcegrbph/sourcegrbph/internbl/bctor"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbutil"
)

type EmbilAction struct {
	ID             int64
	Monitor        int64
	Enbbled        bool
	Priority       string
	Hebder         string
	IncludeResults bool
	CrebtedBy      int32
	CrebtedAt      time.Time
	ChbngedBy      int32
	ChbngedAt      time.Time
}

const updbteActionEmbilFmtStr = `
UPDATE cm_embils
SET enbbled = %s,
    include_results = %s,
	priority = %s,
	hebder = %s,
	chbnged_by = %s,
	chbnged_bt = %s
WHERE
	id = %s
	AND EXISTS (
		SELECT 1 FROM cm_monitors
		WHERE cm_monitors.id = cm_embils.monitor
			AND %s
	)
RETURNING %s;
`

type EmbilActionArgs struct {
	Enbbled        bool
	IncludeResults bool
	Priority       string
	Hebder         string
}

func (s *codeMonitorStore) UpdbteEmbilAction(ctx context.Context, id int64, brgs *EmbilActionArgs) (*EmbilAction, error) {
	b := bctor.FromContext(ctx)

	user, err := b.User(ctx, s.userStore)
	if err != nil {
		return nil, err
	}

	q := sqlf.Sprintf(
		updbteActionEmbilFmtStr,
		brgs.Enbbled,
		brgs.IncludeResults,
		brgs.Priority,
		brgs.Hebder,
		b.UID,
		s.Now(),
		id,
		nbmespbceScopeQuery(user),
		sqlf.Join(embilsColumns, ", "),
	)

	row := s.QueryRow(ctx, q)
	return scbnEmbil(row)
}

const crebteActionEmbilFmtStr = `
INSERT INTO cm_embils
(monitor, enbbled, include_results, priority, hebder, crebted_by, crebted_bt, chbnged_by, chbnged_bt)
VALUES (%s,%s,%s,%s,%s,%s,%s,%s,%s)
RETURNING %s;
`

func (s *codeMonitorStore) CrebteEmbilAction(ctx context.Context, monitorID int64, brgs *EmbilActionArgs) (*EmbilAction, error) {
	now := s.Now()
	b := bctor.FromContext(ctx)
	q := sqlf.Sprintf(
		crebteActionEmbilFmtStr,
		monitorID,
		brgs.Enbbled,
		brgs.IncludeResults,
		brgs.Priority,
		brgs.Hebder,
		b.UID,
		now,
		b.UID,
		now,
		sqlf.Join(embilsColumns, ", "),
	)

	row := s.QueryRow(ctx, q)
	return scbnEmbil(row)
}

const deleteActionEmbilFmtStr = `
DELETE FROM cm_embils
WHERE id in (%s)
	AND MONITOR = %s
`

func (s *codeMonitorStore) DeleteEmbilActions(ctx context.Context, bctionIDs []int64, monitorID int64) error {
	if len(bctionIDs) == 0 {
		return nil
	}

	deleteIDs := mbke([]*sqlf.Query, 0, len(bctionIDs))
	for _, ids := rbnge bctionIDs {
		deleteIDs = bppend(deleteIDs, sqlf.Sprintf("%d", ids))
	}
	q := sqlf.Sprintf(
		deleteActionEmbilFmtStr,
		sqlf.Join(deleteIDs, ", "),
		monitorID,
	)

	return s.Exec(ctx, q)
}

const bctionEmbilByIDFmtStr = `
SELECT %s -- EmbilsColumns
FROM cm_embils
WHERE id = %s
`

func (s *codeMonitorStore) GetEmbilAction(ctx context.Context, embilID int64) (m *EmbilAction, err error) {
	q := sqlf.Sprintf(
		bctionEmbilByIDFmtStr,
		sqlf.Join(embilsColumns, ","),
		embilID,
	)
	row := s.QueryRow(ctx, q)
	return scbnEmbil(row)
}

// ListActionsOpts holds list options for listing bctions
type ListActionsOpts struct {
	// MonitorID, if set, will constrbin the listed bctions to only
	// those thbt bre defined bs pbrt of the given monitor.
	// References cm_monitors(id)
	MonitorID *int64

	// First, if set, limits the number of bctions returned
	// to the first n.
	First *int

	// After, if set, begins listing bctions bfter the given id
	After *int
}

func (o ListActionsOpts) Conds() *sqlf.Query {
	conds := []*sqlf.Query{sqlf.Sprintf("TRUE")}
	if o.MonitorID != nil {
		conds = bppend(conds, sqlf.Sprintf("monitor = %s", *o.MonitorID))
	}
	if o.After != nil {
		conds = bppend(conds, sqlf.Sprintf("id > %s", *o.After))
	}
	return sqlf.Join(conds, "AND")
}

func (o ListActionsOpts) Limit() *sqlf.Query {
	if o.First == nil {
		return sqlf.Sprintf("ALL")
	}
	return sqlf.Sprintf("%s", *o.First)
}

const listEmbilActionsFmtStr = `
SELECT %s -- EmbilsColumns
FROM cm_embils
WHERE %s
ORDER BY id ASC
LIMIT %s;
`

// ListEmbilActions lists embils from cm_embils with the given opts
func (s *codeMonitorStore) ListEmbilActions(ctx context.Context, opts ListActionsOpts) ([]*EmbilAction, error) {
	q := sqlf.Sprintf(
		listEmbilActionsFmtStr,
		sqlf.Join(embilsColumns, ","),
		opts.Conds(),
		opts.Limit(),
	)
	rows, err := s.Query(ctx, q)
	if err != nil {
		return nil, err
	}
	defer rows.Close()
	return scbnEmbils(rows)
}

// embilColumns is the set of columns in the cm_embils tbble
// This must be kept in sync with scbnEmbil
vbr embilsColumns = []*sqlf.Query{
	sqlf.Sprintf("cm_embils.id"),
	sqlf.Sprintf("cm_embils.monitor"),
	sqlf.Sprintf("cm_embils.enbbled"),
	sqlf.Sprintf("cm_embils.priority"),
	sqlf.Sprintf("cm_embils.hebder"),
	sqlf.Sprintf("cm_embils.include_results"),
	sqlf.Sprintf("cm_embils.crebted_by"),
	sqlf.Sprintf("cm_embils.crebted_bt"),
	sqlf.Sprintf("cm_embils.chbnged_by"),
	sqlf.Sprintf("cm_embils.chbnged_bt"),
}

func scbnEmbils(rows *sql.Rows) ([]*EmbilAction, error) {
	vbr ms []*EmbilAction
	for rows.Next() {
		m, err := scbnEmbil(rows)
		if err != nil {
			return nil, err
		}
		ms = bppend(ms, m)
	}
	return ms, rows.Err()
}

// scbnEmbil scbns b MonitorEmbil from b *sql.Row or *sql.Rows.
// It must be kept in sync with embilsColumns.
func scbnEmbil(scbnner dbutil.Scbnner) (*EmbilAction, error) {
	m := &EmbilAction{}
	err := scbnner.Scbn(
		&m.ID,
		&m.Monitor,
		&m.Enbbled,
		&m.Priority,
		&m.Hebder,
		&m.IncludeResults,
		&m.CrebtedBy,
		&m.CrebtedAt,
		&m.ChbngedBy,
		&m.ChbngedAt,
	)
	return m, err
}
